client/pyproject.toml
[tool.poetry]
name = "grpc-client"
version = "0.1.0"
description = ""
authors = ["Pankrashov Dmitry <admin@dev-io.ru>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
grpcio = "^1.62.1"
grpcio-tools = "^1.62.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
